+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Fixed
+- Errors no longer return a code of 0.
 
 ## [1.1.1] - 2015-10-18
 ### Fixed